requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
groq==0.4.2
ibm-watson-machine-learning==1.0.335
python-dotenv==1.0.0
pydantic==2.5.3
